requirements.txt
anyio==3.7.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
contourpy==1.1.1
cryptography==41.0.2
cycler==0.12.0
fonttools==4.43.0
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
mysql-connector-python==8.0.33
numpy==1.26.0
packaging==23.2
Pillow==10.0.0
pycparser==2.21
pyodbc==4.0.39
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tabulate==0.9.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
